requirements.txt
alembic==1.0.0
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
click==6.7
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.12.1
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.4
mccabe==0.6.1
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
psycopg2==2.7.5
py==1.5.4
PyJWT==1.4.2
pylint==2.1.1
pytest==3.7.2
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.11
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
